Trait atspi_client::accessible_ext::AccessibleExt

pub trait AccessibleExt {
    type Error: Error;

    // Required methods
    fn get_application_ext<'a, 'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Self, Self::Error>> + Send + 'async_trait>>
       where Self: Sized + 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait;
    fn get_parent_ext<'a, 'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Self, Self::Error>> + Send + 'async_trait>>
       where Self: Sized + 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait;
    fn get_children_ext<'a, 'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Self>, Self::Error>> + Send + 'async_trait>>
       where Self: Sized + 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait;
    fn get_siblings<'a, 'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Self>, Self::Error>> + Send + 'async_trait>>
       where Self: Sized + 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait;
    fn get_children_indexes<'a, 'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<i32>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait;
    fn get_siblings_before<'a, 'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Self>, Self::Error>> + Send + 'async_trait>>
       where Self: Sized + 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait;
    fn get_siblings_after<'a, 'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Self>, Self::Error>> + Send + 'async_trait>>
       where Self: Sized + 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait;
    fn get_children_caret<'a, 'life0, 'async_trait>(
        &'life0 self,
        after: bool
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Self>, Self::Error>> + Send + 'async_trait>>
       where Self: Sized + 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait;
    fn get_next<'a, 'life0, 'life1, 'async_trait>(
        &'life0 self,
        matcher_args: &'life1 MatcherArgs,
        backward: bool,
        already_visited: &'a mut Vec<ObjectPair>
    ) -> Pin<Box<dyn Future<Output = Result<Option<Self>, Self::Error>> + Send + 'async_trait>>
       where Self: Sized + 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn edges<'a, 'life0, 'async_trait>(
        &'life0 self,
        backward: Option<bool>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Self>, Self::Error>> + Send + 'async_trait>>
       where Self: Sized + 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait;
    fn get_relation_set_ext<'a, 'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<HashMap<RelationType, Vec<Self>>, Self::Error>> + Send + 'async_trait>>
       where Self: Sized + 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait;
    fn match_<'life0, 'life1, 'async_trait>(
        &'life0 self,
        matcher_args: &'life1 MatcherArgs
    ) -> Pin<Box<dyn Future<Output = Result<bool, <Self as AccessibleExt>::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Required Associated Types

type Error: Error

Required Methods

fn get_application_ext<'a, 'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Self, Self::Error>> + Send + 'async_trait>>
where Self: Sized + 'async_trait, 'a: 'async_trait, 'life0: 'async_trait,

fn get_parent_ext<'a, 'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Self, Self::Error>> + Send + 'async_trait>>
where Self: Sized + 'async_trait, 'a: 'async_trait, 'life0: 'async_trait,

fn get_children_ext<'a, 'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<Self>, Self::Error>> + Send + 'async_trait>>
where Self: Sized + 'async_trait, 'a: 'async_trait, 'life0: 'async_trait,

fn get_siblings<'a, 'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<Self>, Self::Error>> + Send + 'async_trait>>
where Self: Sized + 'async_trait, 'a: 'async_trait, 'life0: 'async_trait,

fn get_children_indexes<'a, 'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<i32>, Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'a: 'async_trait, 'life0: 'async_trait,

fn get_siblings_before<'a, 'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<Self>, Self::Error>> + Send + 'async_trait>>
where Self: Sized + 'async_trait, 'a: 'async_trait, 'life0: 'async_trait,

fn get_siblings_after<'a, 'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<Self>, Self::Error>> + Send + 'async_trait>>
where Self: Sized + 'async_trait, 'a: 'async_trait, 'life0: 'async_trait,

fn get_children_caret<'a, 'life0, 'async_trait>( &'life0 self, after: bool ) -> Pin<Box<dyn Future<Output = Result<Vec<Self>, Self::Error>> + Send + 'async_trait>>
where Self: Sized + 'async_trait, 'a: 'async_trait, 'life0: 'async_trait,

fn get_next<'a, 'life0, 'life1, 'async_trait>( &'life0 self, matcher_args: &'life1 MatcherArgs, backward: bool, already_visited: &'a mut Vec<ObjectPair> ) -> Pin<Box<dyn Future<Output = Result<Option<Self>, Self::Error>> + Send + 'async_trait>>
where Self: Sized + 'async_trait, 'a: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn edges<'a, 'life0, 'async_trait>( &'life0 self, backward: Option<bool> ) -> Pin<Box<dyn Future<Output = Result<Vec<Self>, Self::Error>> + Send + 'async_trait>>
where Self: Sized + 'async_trait, 'a: 'async_trait, 'life0: 'async_trait,

Get all edges for a given accessible object. This means: all children, siblings, and parent, in that order. If a direction is specified, then it will only get the appicable matching siblings/children. This also checks if the element supports the text interface, and then checks if the caret position is contained within the string, if it is, then children are also handled by direction.

fn get_relation_set_ext<'a, 'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<HashMap<RelationType, Vec<Self>>, Self::Error>> + Send + 'async_trait>>
where Self: Sized + 'async_trait, 'a: 'async_trait, 'life0: 'async_trait,

fn match_<'life0, 'life1, 'async_trait>( &'life0 self, matcher_args: &'life1 MatcherArgs ) -> Pin<Box<dyn Future<Output = Result<bool, <Self as AccessibleExt>::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,