Trait atspi_client::text_ext::TextBlockingExt

pub trait TextBlockingExt: TextBlockingExtError {
    // Required method
    fn get_all_text(
        &self
    ) -> Result<String, <Self as TextBlockingExtError>::Error>;
}

Required Methods

fn get_all_text(&self) -> Result<String, <Self as TextBlockingExtError>::Error>

Gets the full text within the accessible item.

Errors

This may fail based on the implementation of Text::get_text or TextBlocking::get_text. With the TextProxy and TextProxyBlocking implmentations, this can fail if you ask for an invalid start or end index, or if the DBus method fails to send or receive.